Fix Windows build broken by 6f9a7a5.
Change "int" to "ushort" to avoid the warning:
conversion from 'int' to 'ushort', possible loss of data

Review-URL: https://codereview.appspot.com/305970043

+++ core/arch/emit_utils_shared.c
@@ -5211,7 +5211,7 @@ emit_special_ibl_xfer(dcontext_t *dcontext, byte *pc, generated_code_t *code,
      * spilled in TLS_REG0_SLOT before calling the ibl routine.
      */
     reg_id_t stub_reg = IF_AARCH64_ELSE(SCRATCH_REG0, SCRATCH_REG1);
-    int stub_slot = IF_AARCH64_ELSE(TLS_REG0_SLOT, TLS_REG1_SLOT);
+    ushort stub_slot = IF_AARCH64_ELSE(TLS_REG0_SLOT, TLS_REG1_SLOT);
     IF_X86(size_t len;)
     byte *ibl_tgt = special_ibl_xfer_tgt(dcontext, code, IBL_LINKED, ibl_type);
     bool absolute = !code->thread_shared;
